Adding in integrity (#31)
* Fixed bad indenting

* Removed unfinished implementation of integrity

Was causing issues

* Finished adding structure for integrity

Chat is working

* Cleaning up comments

* Added names in front of chat messages

* Starting to implement integrity

(switching computers)

* Server can send files with integrity to client

(Switching computers again)

* Updating Integrity files

likely need work

* Implementing integrity *Still has issues with CI, CIA, IA

The padding is making it difficult to get same hash value when the receiver decrypts the message. We need to fix it somehow. Maybe padding message before signing and sending it?

* Testing length of message

* Implementing integrity

* Commented out MAC for now

CIA should all be working

* Auto stash before merge of "SarahChatAndClientServer" and "origin/SarahChatAndClientServer"